Use HTTP 404 instead of HTTP 500 during errors
A 500 response code is defined as the following:

```
The HyperText Transfer Protocol (HTTP) 500 Internal Server Error server
error response code indicates that the server encountered an unexpected
condition that prevented it from fulfilling the request.

This error response is a generic "catch-all" response. Usually, this
indicates the server cannot find a better 5xx error code to response.
Sometimes, server administrators log error responses like the 500 status
code with more details about the request to prevent the error from
happening again in the future.
```

https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/500

In all cases in `processor.rb` the conditions are expected. We are
expecting requests that do not have the proper signature/event/etc. In
fact responding with 5XX leaks information that the requester can use.
e.g. The signature is incorrect.

I believe these should be responding with a HTTP 404 as this does not
leak the existence of a webhook endpoint. It tells the requester that
the resource can't be found.

```
The HTTP 404 Not Found response status code indicates that the server
cannot find the requested resource. Links that lead to a 404 page are
often called broken or dead links and can be subject to link rot.

A 404 status code only indicates that the resource is missing: not
whether the absence is temporary or permanent. If a resource is
permanently removed, use the 410 (Gone) status instead.
```

https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/404

Rails will recognize `AbstractController::ActionNotFound` and respond
automatically with HTTP 404. Error messages can still be inspected as
they will be logged.

e.g.
```
Completed 404 Not Found in 2ms (ActiveRecord: 0.0ms | Allocations: 1031)

AbstractController::ActionNotFound (pull_request123 is not a whitelisted GitHub event. See https://developer.github.com/v3/activity/events/types/):
```